refactor: rename ha_client to haclient with src/ layout
- Rename package from ha_client to haclient (matching project name HaClient)
- Move package to standard src/haclient/ layout
- Update all imports across source, tests, examples, and docs
- Update pyproject.toml for new package name and src layout
- Update CI workflow coverage target
- Update AGENTS.md, README.md, and all documentation references
- All 129 tests pass with 95.12% coverage
