+++ src/haclient/domains/__init__.py
@@ -1,4 +1,4 @@
-"""Domain-specific :class:`~ha_client.entity.Entity` subclasses."""
+"""Domain-specific :class:`~haclient.entity.Entity` subclasses."""
 
 from .binary_sensor import BinarySensor
 from .climate import Climate
