+++ src/haclient/domains/binary_sensor.py
