+++ src/haclient/domains/sensor.py
