+++ src/haclient/domains/switch.py
